refactor(cosmic-swingset): remove block-manager
